+++ src/modules/cmdpal/extensionsdk/Microsoft.CmdPal.Extensions.Helpers/ISettingsForm.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Text.Json.Nodes;
+
+namespace Microsoft.CmdPal.Extensions.Helpers;
+
+internal interface ISettingsForm
+{
+    public string ToForm();
+    public void Update(JsonObject payload);
+    public Dictionary<string, object> ToDictionary();
+    public string ToDataIdentifier();
+
+}
